Cache. Кэш память (репозиторий)
