+++ app/src/main/java/com/android/retrofitsampleapp/data/GitHubApi.java
@@ -1,7 +1,7 @@
 package com.android.retrofitsampleapp.data;
 
-import com.android.retrofitsampleapp.domain.GitProjectEntity;
-import com.android.retrofitsampleapp.domain.GitUserEntity;
+import com.android.retrofitsampleapp.domain.Project.GitProjectEntity;
+import com.android.retrofitsampleapp.domain.users.GitUserEntity;
 
 import java.util.List;
