+++ app/src/main/java/com/android/retrofitsampleapp/ui/projects/GitProjectAdapter.java
@@ -8,7 +8,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.android.retrofitsampleapp.R;
-import com.android.retrofitsampleapp.domain.GitProjectEntity;
+import com.android.retrofitsampleapp.domain.Project.GitProjectEntity;
 
 import java.util.ArrayList;
 import java.util.List;
