+++ app/src/main/res/layout/activity_users.xml
@@ -10,6 +10,7 @@
         android:id="@+id/load_button"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:padding="8dp"
         android:text="Load" />
 
     <FrameLayout
